refactor: consolidate storage service interfaces
